Allow using fields inside the badge
The field values are exposed under fields.(a|b).NAME.
